Feature/back/corehr (#65)
* feat(back): incidents app e infra layer implemented

* feat(back): corehr domain

Domain implemented.

* feat(back): corehr application

capa de aplicacion y tests implementada

* feat(back): corehr infrastructure

* refactor(back):  refactor module incidents

se migro la logica de TenantUser provisionalmente escrita en el modulo de incidencias al modulo de corehr, modulo al cual pertenece naturalmente.

Y se removio la anotacion "@Service" erroneamente definida en la capa de application.

* refactor(back): modulo corehr e incidents

Alineación con OpenAPI:

discrepancia de validación HTTP: MethodArgumentNotValidException ahora devuelve 400 Bad Request, alineado con errores de request inválido del contrato OpenAPI.

Archivo modificado: [CoreHrGlobalExceptionHandler.java](C:/ProyectosGit/Trazzo/back/src/main/java/trazzo/back/corehr/infrastructure/adapters/in/web/CoreHrGlobalExceptionHandler.java).

Integración de Módulos (CoreHR ↔ Incidents/Reports)

incidents ya usa el puerto público TenantUserPort; no observé acceso directo a repositorios/base de datos de CoreHR.

* refactor(corehr): auditoría arquitectónica, persistencia robusta y alineación OpenAPI

Resumen de Cambios
Refactorización profunda del módulo CoreHR para resolver deuda técnica crítica, estandarizar tipos de datos (Long), alinear la capa de adaptadores con la fuente de verdad (OpenAPI) y blindar la lógica en la capa de Dominio.

1. Arquitectura Hexagonal y Dominio

Estandarización de Identificadores: Se refactorizó TenantUserPort para operar estrictamente con Long. Se eliminaron casteos innecesarios (String.valueOf) en los consumidores.

Persistencia JDBC Precisa: Reescritura total de TenantUserJdbcAdapter para consultar correctamente desde tenant_user (JOIN con users/persons), mapeando atributos omitidos como phone y garantizando la resolución del state.

Dominio Rico en Asistencias (Attendance.java): Se reemplazó la recreación del agregado por comportamiento real. Ahora updateCheckIn y updateCheckOut disparan touch() internamente, asegurando la consistencia temporal.

Integridad Biométrica (UserBiometria.java): Extensión del modelo de dominio para incorporar y persistir templateCifrado y llaveCifrado.

2. Alineación Contractual (OpenAPI)

Filtros Funcionales: Se conectaron los parámetros de consulta ignorados (branch_id, area_id, departamento_id, tenant_user_id, rangos y estado) directamente al AttendanceRepositoryAdapter. El parámetro scope se mantiene como passthrough.

Contratos de Contacto: Corrección de TenantContactResult y TenantContactResponse para exponer el atributo phone (previamente en nulo) y asegurar la consistencia del ID como Long.

Privacidad de Biometría: El payload cifrado ahora se persiste correctamente mediante el EnrollService y su Mapper, asegurando que los datos sensibles jamás se expongan en las respuestas GET.

3. Integración Inter-Módulos (Modulith)

Módulo reports: Alineación end-to-end del tenantUserId a Long en puertos de salida, dominio (MonthlyClosureDetail) y adaptadores. Se corrigió un bug de truncamiento en CoreHrAttendanceSummaryJdbcAdapter aplicando rs.getLong("id").

Módulo incidents: Actualización en IncidentService para consumir el nuevo flujo, agregando tolerancia de parseo para mantener retrocompatibilidad con IDs legacy no numéricos.

Límites Arquitectónicos: Se validó que la comunicación cross-módulo se mantenga 100% aislada a través de Puertos de integración.

4. Pruebas y Quality Gate (SonarQube)

Cobertura Ampliada: Se implementaron pruebas exhaustivas para TenantUserJdbcAdapter, TenantContactService, flujos de cifrado biométrico, y casos de borde en correcciones de asistencia.

Estado de la Suite: 1056 tests ejecutados | 0 fallos | 0 errores.

5. Compilación y Deuda Técnica

Build exitoso y limpio al ejecutar mvn clean verify.

Nota Técnica: Se documenta para el backlog un Deprecation Warning en SecurityConfig.java relacionado con AntPathRequestMatcher (no bloqueante para esta entrega).

* refactor(back): duplication & tests

* test(back): coverage completo

* test(back): correcion test

error en ci backend tests:

The test asserted strict equality (isEqualTo) between the two, but nanosecond drift between T1 and T2 occasionally caused failure.

Fix: Replaced strict equality with nanosecond-tolerant assertion

This truncates nanoseconds before comparing, so differences < 1 millisecond are ignored entirely.

+++ back/src/main/java/trazzo/back/reports/application/dto/result/MonthlyClosureDetailResult.java
@@ -6,7 +6,7 @@ import java.util.UUID;
 public record MonthlyClosureDetailResult(
         UUID id,
         UUID monthClosureId,
-        Integer tenantUserId,
+        Long tenantUserId,
         String tenantUserFullName,
         String tenantUserDocument,
         String departmentName,
